Update WxWidgets repository
This is in preparation to make it easy to compile with VS2015.

Also there's notables changes in wxWidgets that we might find very useful,
such as being able to create OpenGL debug contexts.

The change in .gitmodules might make it necessary for people to run
"git submodule sync" after pulling these changes