Emu: per game/ELF file configuration and cache functionality
Adds a per game and ELF file configuration and cache functionality to the emulator.
The SHA-1 hash of the last 20 characters of the file path or game path is used to
generate a relative path, the original filename is appened to it.

Usage:
  fs::get_data_dir(file path, title id = optimal)

Returns the absolute path where the configuration or cache for the current ELF file
or game is stored.