VS build: delete QT DLLs from bin directory
Commit 6cfb184b1ec798eb209272a404040e7a582516a7 added these files when
implementing a very large code refactoring. They seems to be unecessary
since there's a couple of VS post-build scripts that copy the files from
the user local QT installation to the bin directory every time the rpcs3
project is built. So their presence in the repository only makes cloning
it a little slower.