gui: Remove fallback font from YoRHa theme
Starting with QT 5.13, having both fonts causes stylesheets to use the fallback font even if the first font is present, but uses the first font when there's no fallback font.
The fallback font (Arial) can safely be removed as the default one is 'good enough' for the intended purpose of having a fallback.