Lightweight putllc() for non-TSX if no data changed (v3)
In case SPU did not change any cacheline data between getllar() and putllc()
just do nothing and report success. If we start a writer locked data transfer
in this situation the only meaningful outcome would be that PPU has already
overwritten main memory. This would be wasted time and resources.