FIX: Avoid possible non-null terminated truncation in NP
strncpy will truncate the string when the count value is reached, but it will truncate it without adding any null terminator.  This makes it trivial for a malicious actor to trigger a buffer overrun (by just providing a string that exceeds the buffer length).  Changing this to strncpy_s and providing the destination size as the dest buffer size, but the source count as dest buffer size -1.  This ensures that even in the event of truncation we will get a null terminated buffer.