HIDAPI: Return to mainline
Changes from the custom repository back to the main libusb repository, and the cmake-support branch.
This branch will be merged into the master branch at somepoint in the near future.
At this time rpcs3 should also update to reference the master branch.

The only change required within rpcs3 code is the replacement of references to hid_write_control with
the hid_write() function.